+++ src/main/java/com/barricador/client/internal/MetricsBuffer.java
@@ -1,4 +1,4 @@
-package com.barricator.client.internal;
+package com.barricador.client.internal;
 
 import java.util.ArrayList;
 import java.util.List;